feat: Add new llama model
