+++ presets/workspace/models/llama3/README.md
@@ -1,0 +1,13 @@
+## Supported Models
+
+| Model name           | Model source                                                                 | Sample workspace                                                                              | Kubernetes Workload | Distributed inference |
+|----------------------|------------------------------------------------------------------------------|----------------------------------------------------------------------------------------------|---------------------|-----------------------|
+| llama3.1-8b-instruct | [llama3.1-8b-instruct](https://huggingface.co/meta-llama/Llama-3.1-8B-Instruct) | [link](../../../../examples/inference/kaito_workspace_llama_3_1_8b_instruct.yaml)           | Deployment          | false                 |
+
+## Image Source
+
+- **Public**: Kaito maintainers manage the lifecycle of the inference service images that contain model weights. The images are available in Microsoft Container Registry (MCR).
+
+## Usage
+
+See [document](../../../../docs/inference/README.md).
